test(query-actions): move drill-detail test to own module, reframe scope
Retarget the #28562 filter test: it only characterizes
_get_drill_detail's pass-through of QueryObject.filter (which the
function never touches), not a reproduction of the reported bug, and
move it out of the currency-detection test module into its own file
per reviewer feedback.
